django_backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
Pillow==10.2.0
geopy==2.4.1
django-cors-headers==4.3.1
python-decouple==3.8
python-dotenv